==> notebooks/03_merge_1_&_2.ipynb <==
import pandas as pd
import numpy as np
import dask.dataframe as dd
import os
import sys
import missingno as msno
import matplotlib.pyplot as plt
import seaborn as sns
from tqdm import tqdm, tqdm_notebook
from collections import Counter
tqdm.pandas(tqdm_notebook)
import gc

%%time
train1 = pd.read_csv('train_processed.csv')
train2 = pd.read_csv('train_processed_2.csv')
test1 = pd.read_csv('test_processed.csv')
test2 = pd.read_csv('test_processed_2.csv')

train1.shape,train2.shape,test1.shape,test2.shape

train1.head()

used_columns = ['card_id'] + [x for x in train2.columns if x not in train1.columns ]

%%time
train = train1.merge(train2[used_columns],how = 'left',on = 'card_id')
test = test1.merge(test2[used_columns],how = 'left',on = 'card_id')

train.shape,test.shape

train.to_csv('train_merged.csv',index=False)
test.to_csv('test_merged.csv',index=False)
